requirements.txt
asgi-redis==0.13.1
asgiref==0.13.3
autobahn==0.14.1
channels==0.14.2
daphne==0.12.1
dj-database-url==0.4.1
Django==1.9.7
django-extensions==1.6.7
django-model-utils==2.5
gunicorn==19.6.0
msgpack-python==0.4.7
psycopg2==2.6.1
redis==2.10.5
six==1.10.0
Twisted==16.2.0
txaio==2.5.1
zope.interface==4.2.0
django-jet==0.1.4
google-api-python-client==1.5.1
